Fix Failed Test When No Block & Label Width Is 12
The ordering of the Object.assign statement, combined with the dynamic
property name, resulted in the same key being assigned twice, once for
true and once for false. Either the "block" condition or the
"labelWidth" condition needs to win to break the tie. This change makes
it so that the "labelWidth" condition overrides the "block" condition.

This change produces the expected output for
`{labelWidth: { xs: 12 }, block: false }`. (col-xs-12)