Convert Error To Component & Add Error Body
Convert the Error to a component because enzyme currently has issues
with stateless components.

Add Error Body component since React's CSS Transition Group does not
render the text, in the Error Component, until a event happens. In order
to avoid this new component, for the body, was create to test how the
errors renders.