Add To Timepicker Component Testing (#34)
* Add TimeFormatter For Time Formatting
* Add Object Freeze To Default Props & Type
* Refactor Time Picker To Be More Code Readable …

Refactor the Time Picker code to be more readable because it has
become to complex with having the formula inside its popup. Which will
make testing both the TimePicker and popup hard to do.

So the plan for the TimePicker is as follows. Split the time when it
comes into the TimePicker, from Frig, and send into a new class
called TimeFormatter. Where the time will be formatted for the
TimePickerPopup to use.

If the user changes the time in the Time Picker Popup the time will
be combine and sent back to the TimeFormatter, in TimePicker,
before sending to the TimePicker and to Frig in general.

* Add Package Cloner To Clone Objects Deeply …

Many of our tests were passing by coincidence because we relied
on Object.assign to deeply copy/merge objects, but it only goes
one level down (shallow).